Validate max instances config
