+++ packages/netsuite-adapter/src/adapter_creator.ts
@@ -19,7 +19,7 @@ import { logger } from '@salto-io/logging'
 import _ from 'lodash'
 import { SdkDownloadService } from '@salto-io/suitecloud-cli'
 import Bottleneck from 'bottleneck'
-import { CLIENT_CONFIG, CONFIG, configType, DEFAULT_CONCURRENCY, DEFAULT_MAX_INSTANCES_VALUE, NetsuiteConfig, validateDeployParams, validateFetchConfig, validateSuiteAppClientParams } from './config'
+import { CONFIG, configType, DEFAULT_CONCURRENCY, DEFAULT_MAX_INSTANCES_VALUE, NetsuiteConfig, validateClientConfig, validateDeployParams, validateFetchConfig, validateSuiteAppClientParams } from './config'
 import { NETSUITE } from './constants'
 import { validateFetchParameters, convertToQueryParams, validateNetsuiteQueryParameters, validateArrayOfStrings, validatePlainObject, FETCH_PARAMS } from './query'
 import { Credentials, isSdfCredentialsOnly, isSuiteAppCredentials, toCredentialsAccountId } from './client/credentials'
@@ -31,8 +31,6 @@ import NetsuiteAdapter from './adapter'
 const log = logger(module)
 
 const configID = new ElemID(NETSUITE)
-// Taken from https://github.com/salto-io/netsuite-suitecloud-sdk/blob/e009e0eefcd918635353d093be6a6c2222d223b8/packages/node-cli/src/validation/InteractiveAnswersValidator.js#L27
-const SUITEAPP_ID_FORMAT_REGEX = /^[a-z0-9]+(\.[a-z0-9]+){2}$/
 
 // The SuiteApp fields are commented out until we will be ready to expose them to the user
 export const defaultCredentialsType = new ObjectType({
@@ -73,14 +71,6 @@ export const defaultCredentialsType = new ObjectType({
   annotations: {},
 })
 
-const validateInstalledSuiteApps = (installedSuiteApps: unknown): void => {
-  validateArrayOfStrings(installedSuiteApps, [CONFIG.client, CLIENT_CONFIG.installedSuiteApps])
-  const invalidValues = installedSuiteApps.filter(id => !SUITEAPP_ID_FORMAT_REGEX.test(id))
-  if (invalidValues.length !== 0) {
-    throw new Error(`${CLIENT_CONFIG.installedSuiteApps} values should contain only lowercase characters or numbers and exactly two dots (such as com.saltoio.salto). The following values are invalid: ${invalidValues.join(', ')}`)
-  }
-}
-
 const validateRegularExpressions = (regularExpressions: string[]): void => {
   const invalidRegularExpressions = regularExpressions
     .filter(strRegex => !regex.isValidRegex(strRegex))
@@ -110,20 +100,8 @@ function validateConfig(config: Record<string, unknown>): asserts config is Nets
     validateArrayOfStrings(typesToSkip, CONFIG.typesToSkip)
   }
 
-  if (client !== undefined) {
-    validatePlainObject(client, CONFIG.client)
-    const {
-      fetchAllTypesAtOnce,
-      installedSuiteApps,
-    } = _.pick(client, Object.values(CLIENT_CONFIG))
-
-    if (fetchAllTypesAtOnce && fetchTarget !== undefined) {
-      log.warn(`${CLIENT_CONFIG.fetchAllTypesAtOnce} is not supported with ${CONFIG.fetchTarget}. Ignoring ${CLIENT_CONFIG.fetchAllTypesAtOnce}`)
-      client[CLIENT_CONFIG.fetchAllTypesAtOnce] = false
-    }
-    if (installedSuiteApps !== undefined) {
-      validateInstalledSuiteApps(installedSuiteApps)
-    }
+  if (client && client !== undefined) {
+    validateClientConfig(client, fetchTarget !== undefined)
   }
 
   if (fetchTarget !== undefined) {
